+++ src/utils/helper-functions.ts
@@ -1,3 +1,11 @@
+/**
+ * Convert special characters to their ASCII number values.
+ * 
+ * @param str Input string to be processed.
+ * @returns Input string with special characters replaced with 
+ *          their ASCII values.
+ */
+
 export const specialToASCII = (str: string): string => {
     let res = '';
     for(let i = 0; i < str.length; i++) {