+++ ui-tests/tests/widget.test.ts
@@ -645,3 +645,51 @@ test('test error if edit connection with existing name', async ({ page }) => {
 
     await expect(page.locator('.user-error-message')).toContainText("A connection named 'default' already exists in your connections file");
 });
+
+const specialAliases = [
+  {alias: 'db', id: 'id'},
+  {alias: 'db with space', id: 'db_with_space'},
+  {alias: 'db?', id: 'db63'},
+  {alias: 'db://', id: 'db584747'},
+  {alias: 'db! !', id: 'db33_33'},
+  {alias: ';db', id: '59db'},
+  {alias: ' db', id: '_db'},
+  { 
+    alias: '!@#$%^&*()-_[]{}|;:?<>,./', 
+    id: '33643536379438424041459591931231251245958636062444647'
+  },
+]
+
+for (const {alias, id} of specialAliases) {
+  test(`test delete button for aliases with special chars`, async ({ page }) => {
+    // create a new connection
+    await page.locator('#createNewConnection').click();
+    await page.locator('#connectionName').fill(alias);
+    await page.locator('#createConnectionFormButton').click();
+
+    // delete connection with special characters in alias
+    await page.locator(`#deleteConnBtn_${id}`).click();
+    await page.locator('#deleteConnectionButton').click();
+
+    expect(page.locator('#connectionsButtonsContainer')).toBeEmpty();
+
+  })
+}
+
+for (const {alias, id} of specialAliases) {
+  test(`test connect button for aliases with special chars`, async ({ page }) => {
+    // create default connection
+    await createDefaultConnection(page);
+
+    // create a new connection
+    await page.locator('#createNewConnectiton').click();
+    await page.locator('#connectionName').fill(alias);
+    await page.locator('#createConnectionFormButton').click();
+
+    // connect to default, then connect back to new connection
+    await page.locator('#connBtn_default').click();
+    await page.locator(`#connBtn_${id}`).click();
+
+    await expect(page.locator('.connection-name')).toContainText(alias);
+  })
+}